Extracted player controls components - play-pause, skip, rewind, like, speed buttons
