+++ src/components/common/player/components/Player.js
@@ -7,7 +7,7 @@ import styled from 'styled-components';
 import EpisodesImages from './episodes-images/EpisodesImages';
 import usePlayer from './usePlayer';
 import EpisodeCover from './EpisodeCover';
-import Button from './Button';
+import PlayerControls from './PlayerControls';
 import ProgressBar from './ProgressBar';
 
 const Wrapper = styled(View)`
@@ -29,14 +29,6 @@ const EpisodeTitleText = styled(Text)`
 const AuthorText = styled(EpisodeTitleText)`
   font-size: ${({ theme }) => theme.metrics.largeSize}px;
   margin: ${({ theme }) => theme.metrics.smallSize}px;
-`;
-
-const ButtonsWrapper = styled(View)`
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-around;
-  align-items: center;
 `;
 
 const Player = ({ route }) => {
@@ -63,43 +55,10 @@ const Player = ({ route }) => {
         {' '}
       </AuthorText>
       <ProgressBar />
-      <ButtonsWrapper>
-        <Button
-          onPress={() => console.log('Speed up pressed')}
-          name="speedometer"
-          size={35}
-        />
-        <Button
-          onPress={() => console.log('Rewind back pressed')}
-          name="skip-previous-circle"
-          size={55}
-        />
-        <Button
-          onPress={() => TrackPlayer.play()}
-          name="play-circle"
-          size={70}
-        />
-        <Button
-          onPress={() => TrackPlayer.pause()}
-          name="pause-circle"
-          size={70}
-        />
-        <Button
-          onPress={() => console.log('Skip button pressed')}
-          name="skip-next-circle"
-          size={55}
-        />
-        <Button
-          onPress={() => console.log('Like is pressed')}
-          name="heart-outline"
-          size={35}
-        />
-      </ButtonsWrapper>
+
+      <PlayerControls />
     </Wrapper>
   );
 };
-
-// speedometer
-// surround-sound-2-0
 
 export default Player;
