+++ src/components/common/player/components/episode-list/EpisodeListItem.js
@@ -44,7 +44,10 @@ const EpisodeListItem = ({
       source={{ uri: image }}
     />
     <TextWrapper>
-      <EpisodeText>{episodeTitle}</EpisodeText>
+      <EpisodeText>
+        {episodeTitle}
+        {' '}
+      </EpisodeText>
       <AuthorText>{author}</AuthorText>
     </TextWrapper>
   </Wrapper>
